# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI Train API.",
        "title": "AI Train API",
        "contact": {},
        "license": {},
        "version": "0.2"
    },
    "host": "localhost:38080",
    "basePath": "/",
    "paths": {
...
        "/v1/job/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a running job deployment in user namespace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job"
                ],
                "summary": "Delete a running job deployment in user namespace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "job uuid, eg: 131ba8a9-b60b-44f9-83b5-46590f756f41",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericOKResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/job/launch": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a course deployment in kubernetes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job"
                ],
                "summary": "Create a course deployment in kubernetes",
                "parameters": [
                    {
                        "description": "course want to launch",
                        "name": "launch_course",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.LaunchCourseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.LaunchCourseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/job/list": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List all running course deployment for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job"
                ],
                "summary": "List all running course deployment for a user",
                "parameters": [
                    {
                        "description": "search user's job",
                        "name": "list_user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.OauthUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.JobListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docs.GenericErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean",
                    "format": "bool",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "example": "error response message"
                }
            }
        },
        "docs.GenericOKResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean",
                    "format": "bool",
                    "example": false
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "example": "response message"
                }
            }
        },
...
        "docs.JobInfo": {
            "type": "object",
            "properties": {
                "course_id": {
                    "type": "string",
                    "example": "b86b2893-b876-45c2-a3f6-5e099c15d638"
                },
                "dataset": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "cifar-10",
                        "mnist"
                    ]
                },
                "gpu": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "id": {
                    "type": "string",
                    "example": "49a31009-7d1b-4ff2-badd-e8c717e2256c"
                },
                "image": {
                    "type": "string",
                    "example": "nvidia/caffe:latest"
                },
                "introduction": {
                    "type": "string",
                    "example": "markdown text with escape"
                },
                "level": {
                    "type": "string",
                    "example": "basic"
                },
                "name": {
                    "type": "string",
                    "example": "mage process"
                },
                "service": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/docs.SVCLabelValue"
                    }
                },
                "startAt": {
                    "type": "string",
                    "example": "2018-06-25T09:24:38Z"
                },
                "status": {
                    "type": "string",
                    "example": "Ready"
                }
            }
        },
        "docs.JobListResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean",
                    "format": "bool",
                    "example": false
                },
                "jobs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/docs.JobInfo"
                    }
                }
            }
        },
        "docs.JobStatus": {
            "type": "object",
            "properties": {
                "job_id": {
                    "type": "string",
                    "example": "5ab02011-9ab7-40c3-b691-d335f93a12ee"
                },
                "ready": {
                    "type": "boolean",
                    "format": "bool",
                    "example": false
                },
                "status": {
                    "type": "string",
                    "example": "Created"
                }
            }
        },
        "docs.LaunchCourseRequest": {
            "type": "object",
            "properties": {
                "course_id": {
                    "type": "string",
                    "example": "5ab02011-9ab7-40c3-b691-d335f93a12ee"
                },
                "user": {
                    "type": "string",
                    "example": "ogre@gamil.com"
                }
            }
        },
        "docs.LaunchCourseResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean",
                    "format": "bool",
                    "example": false
                },
                "job": {
                    "type": "object",
                    "$ref": "#/definitions/docs.JobStatus"
                }
            }
        },
...
        "docs.OauthUser": {
            "type": "object",
            "properties": {
                "user": {
                    "type": "string",
                    "example": "ogre@gamil.com"
                }
            }
        },
...
        "docs.SVCLabelValue": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string",
                    "example": "jupyter"
                },
                "value": {
                    "type": "string",
                    "example": "http://140.110.5.22:30010"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}